requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
requests>=2.28.0
python-dotenv>=1.0.0
google-ads
google-auth-oauthlib
